Fix probing on the alpha.  It still causes the alpha to panic during attach.
Former-commit-id: 2b7fa42425f8fbd23ab2f3ce860a82c16cfb9bbc

+++ sys/dev/asr/asr.c
@@ -1452,7 +1452,7 @@ define|#
 directive|define
 name|PROBE_SET
 parameter_list|()
-value|u_long id = (pci_get_device(tag)<<16)|pci_get_vendor(tag)
+value|u_int32_t id = (pci_get_device(tag)<<16)|pci_get_vendor(tag)
 end_define
 
 begin_define
